requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-genai
numpy
Pillow
unidecode
scipy
scikit-image
